# vulkan: optimizations for deepseek prompt processing